feat:  clean up example folder
