+++ examples/vanilla/package.json
