+++ examples/vanilla/public/.vite/manifest.json
