+++ examples/vanilla/public/assets/main-w40geAFS.js
