+++ examples/vanilla/resources/bootstrap/ssr/main.js
@@ -1,0 +1,1 @@
+
