+++ examples/vanilla/templates/index.html.j2
@@ -1,5 +1,5 @@
 <!DOCTYPE html>
-<html class="h-full">
+<html>
   <head>
     <meta charset="utf-8" />
     <!--IE compatibility-->
@@ -10,9 +10,9 @@
     />
   </head>
 
-  <body class="font-sans leading-none text-gray-700 antialiased">
+  <body>
     <div id="app"></div>
     {{ vite_hmr() }}
-    {{ vite('main.ts') }}
+    {{ vite('resources/main.ts') }}
   </body>
 </html>
