+++ examples/vanilla/tsconfig.json
@@ -1,15 +1,15 @@
 {
   "compilerOptions": {
-    "target": "ESNext",
+    "target": "ES2020",
     "useDefineForClassFields": true,
-    "module": "ESNext",
-    "moduleResolution": "node",
+    "module": "ES2020",
+    "moduleResolution": "bundler",
     "strict": true,
     "jsx": "preserve",
     "resolveJsonModule": true,
     "isolatedModules": true,
     "esModuleInterop": true,
-    "lib": ["ESNext", "DOM"],
+    "lib": ["ES2020", "DOM", "DOM.Iterable"],
     "skipLibCheck": true,
     "noEmit": true,
     "composite": true,
@@ -18,14 +18,13 @@
     "paths": {
       "@/*": ["./*"]
     },
-    "types": ["vite/client", "node"]
+    "types": ["vite/client"]
   },
   "include": [
+    
     "**/*.d.ts",
     "resources/**/*.js" ,
     "resources/**/*.ts" ,
-
-
     "vite.config.ts"
   ]
 }
