+++ examples/vanilla/vite.config.ts
